Further restructure file hierarchy
