+++ bubblemachine-vite-react/src/App.jsx
@@ -1,7 +1,7 @@
 // src/App.jsx
 import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Layout from "./pages/Layout";
+import Layout from "./components/layout/Layout";
 import HomePage from "./pages/HomePage";
 import WaveformVis from "./components/waveform/WaveformVis";
 import ErrorPage from "./pages/ErrorPage";
