+++ bubblemachine-vite-react/src/components/layout/Header.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { AppBar, Toolbar, Typography, Box } from "@mui/material";
-import logo from "../assets/BubbleMachine_Transparent.png"; // Adjust the path if necessary
+import logo from "../../assets/BubbleMachine_Transparent.png";
+
 const Header = () => {
   return (
     <AppBar position="static">
