+++ bubblemachine-vite-react/src/components/layout/Layout.jsx
@@ -1,7 +1,7 @@
 // src/pages/Layout.jsx
 import React from "react";
-import Header from "../components/layout/Header";
-import Footer from "../components/layout/Footer";
+import Header from "./Header";
+import Footer from "./Footer";
 import { Outlet, Link } from "react-router-dom";
 
 export default function Layout() {
