+++ bubblemachine-vite-react/src/components/table/ZTable.jsx
@@ -6,7 +6,7 @@ import {
 import { Box, Button, IconButton, Tooltip } from '@mui/material';
 import EditIcon from '@mui/icons-material/Edit';
 import DeleteIcon from '@mui/icons-material/Delete';
-import useBubbleStore from '../../state';
+import useBubbleStore from '../zustand/bubbleStore';
 
 const BubbleTable = () => {
     const [validationErrors, setValidationErrors] = useState({});
